Services/GrandPrixService/requirements.txt
fastapi
gunicorn
pydantic==1.10.16
uvicorn
#uvloop
httptools
firebase-admin
google-cloud-bigquery
google-cloud-firestore>=1.6.2
google-cloud-storage
google-cloud-pubsub
pydantic[email]
pytz
httpx
requests
click==8.1.3
httplib2
trio